use new useDataSourceColumn in nodes
