+++ front/components/data_source_view/DataSourceViewTable.tsx
@@ -1,48 +1,22 @@
-import {
-  Checkbox,
-  DataTable,
-  FolderIcon,
-  Hoverable,
-  ScrollableDataTable,
-  Spinner,
-} from "@dust-tt/sparkle";
-import type { ColumnDef } from "@tanstack/react-table";
-import { useMemo } from "react";
+import { FolderIcon, ScrollableDataTable, Spinner } from "@dust-tt/sparkle";
 
 import { useAgentBuilderContext } from "@app/components/agent_builder/AgentBuilderContext";
 import { useDataSourceBuilderContext } from "@app/components/data_source_view/context/DataSourceBuilderContext";
-import type { NavigationHistoryEntryType } from "@app/components/data_source_view/context/types";
 import {
   findCategoryFromNavigationHistory,
   findSpaceFromNavigationHistory,
 } from "@app/components/data_source_view/context/utils";
+import type { DataSourceRowData } from "@app/components/data_source_view/hooks/useDataSourceColumns";
+import { useDataSourceColumns } from "@app/components/data_source_view/hooks/useDataSourceColumns";
 import { useTheme } from "@app/components/sparkle/ThemeContext";
 import { CONNECTOR_CONFIGURATIONS } from "@app/lib/connector_providers";
 import { CATEGORY_DETAILS } from "@app/lib/spaces";
 import { useSpaceDataSourceViews } from "@app/lib/swr/spaces";
-import type { DataSourceViewType } from "@app/types";
-
-type RowData = {
-  id: string;
-  title: string;
-  onClick?: () => void;
-  icon?: React.ComponentType;
-  dataSourceView: DataSourceViewType;
-};
 
 export function DataSourceViewTable() {
   const { owner } = useAgentBuilderContext();
-  const {
-    navigationHistory,
-    isCurrentNavigationEntrySelected,
-    selectCurrentNavigationEntry,
-    removeCurrentNavigationEntry,
-    isRowSelected,
-    isRowSelectable,
-    selectNode,
-    removeNode,
-    setDataSourceViewEntry,
-  } = useDataSourceBuilderContext();
+  const { navigationHistory, setDataSourceViewEntry } =
+    useDataSourceBuilderContext();
   const space = findSpaceFromNavigationHistory(navigationHistory);
   const { isDark } = useTheme();
 
@@ -54,89 +28,8 @@ export function DataSourceViewTable() {
       spaceId: space?.sId ?? "",
     });
 
-  const columns: ColumnDef<RowData>[] = useMemo(
-    () => [
-      {
-        id: "select",
-        enableSorting: false,
-        enableHiding: false,
-        header: () => {
-          const selectionState = isCurrentNavigationEntrySelected();
-          return (
-            <Checkbox
-              key={`header-${selectionState}`}
-              size="xs"
-              checked={selectionState}
-              disabled={!isRowSelectable()}
-              onClick={(event) => event.stopPropagation()}
-              onCheckedChange={(state) => {
-                // When clicking a partial checkbox, select all
-                if (selectionState === "partial" || state) {
-                  selectCurrentNavigationEntry();
-                } else {
-                  removeCurrentNavigationEntry();
-                }
-              }}
-            />
-          );
-        },
-        cell: ({ row }) => {
-          const selectionState = isRowSelected(row.original.id);
-
-          return (
-            <div className="flex h-full w-full items-center">
-              <Checkbox
-                key={`${row.original.id}-${selectionState}`}
-                size="xs"
-                checked={selectionState}
-                disabled={!isRowSelectable(row.original.id)}
-                onClick={(event) => event.stopPropagation()}
-                onCheckedChange={(state) => {
-                  // When clicking a partial checkbox, select all
-                  const item: NavigationHistoryEntryType = {
-                    type: "data_source",
-                    dataSourceView: row.original.dataSourceView,
-                  };
-                  if (selectionState === "partial" || state) {
-                    selectNode(item);
-                  } else {
-                    removeNode(item);
-                  }
-                }}
-              />
-            </div>
-          );
-        },
-        meta: {
-          sizeRatio: 5,
-        },
-      },
-      {
-        accessorKey: "title",
-        id: "name",
-        header: "Name",
-        cell: ({ row }) => (
-          <DataTable.CellContent icon={row.original.icon}>
-            <Hoverable>{row.original.title}</Hoverable>
-          </DataTable.CellContent>
-        ),
-        meta: {
-          sizeRatio: 70,
-        },
-      },
-    ],
-    [
-      isCurrentNavigationEntrySelected,
-      isRowSelectable,
-      isRowSelected,
-      removeCurrentNavigationEntry,
-      removeNode,
-      selectCurrentNavigationEntry,
-      selectNode,
-    ]
-  );
-
-  const rows: RowData[] = spaceDataSourceViews.map((dsv) => {
+  const columns = useDataSourceColumns();
+  const rows: DataSourceRowData[] = spaceDataSourceViews.map((dsv) => {
     const connectorProvider = dsv.dataSource.connectorProvider
       ? CONNECTOR_CONFIGURATIONS[dsv.dataSource.connectorProvider]
       : null;
@@ -158,8 +51,11 @@ export function DataSourceViewTable() {
       id: dsv.sId,
       title,
       onClick: () => setDataSourceViewEntry(dsv),
-      dataSourceView: dsv,
       icon,
+      entry: {
+        type: "data_source",
+        dataSourceView: dsv,
+      },
     };
   });
